+++ M2_Jobs/src/objects/Voluntario.java
@@ -1,12 +1,12 @@
 package objects;
 
 public class Voluntario extends Persona{
-	public Voluntario(int saldo) {
+	public Voluntario(double saldo) {
 		this.saldo = 0;
 	}
 
 	@Override
-	public void setSaldo(int saldo) {
+	public void setSaldo(double saldo) {
 		this.saldo = 0;
 		System.out.println("Los voluntarios no cobran!");
 	}
